Init: Initial commit
Added layout, seo and header components. Updated index and about us page.

Author: Menlik Getachew menlikgetachew@gmail.com
