--- gatsby-browser.js
@@ -1,1 +1,0 @@
-import "./src/styles/global.css"
