+++ package.json
@@ -13,19 +13,9 @@
     "serve": "gatsby serve",
     "clean": "gatsby clean"
   },
-  "dependencies": {
-    "autoprefixer": "^10.4.16",
+ "dependencies": {
     "gatsby": "^5.13.0",
-    "gatsby-plugin-google-gtag": "^5.13.0",
-    "gatsby-plugin-image": "^3.13.0",
-    "gatsby-plugin-postcss": "^6.13.0",
-    "gatsby-plugin-sharp": "^5.13.0",
-    "gatsby-plugin-sitemap": "^6.13.0",
-    "gatsby-source-filesystem": "^5.13.0",
-    "gatsby-transformer-sharp": "^5.13.0",
-    "postcss": "^8.4.32",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "tailwindcss": "^3.4.0"
+    "react-dom": "^18.2.0"
   }
 }
