+++ src/components/header.module.css
@@ -1,0 +1,24 @@
+.site-title {
+    font-size: 2.5rem;
+    color: gray;
+    font-weight: 700;
+    margin: 3rem 0;
+}
+
+.nav-links {
+    display: flex;
+    list-style: none;
+    padding-left: 0;
+}
+
+.nav-link-item {
+    padding-right: 2rem;
+}
+
+.nav-link-text {
+    margin: auto;
+    max-width: 500px;
+    font-family: sans-serif;
+    font-size: 1.5rem;
+    color: black;
+}
