+++ src/components/layout.js
@@ -1,0 +1,20 @@
+import * as React from 'react'
+import Header from './header'
+import {
+    container,
+    heading
+} from "./layout.module.css"
+
+const Layout = ({ pageTitle, children }) => {
+  return (
+    <div className={container}>
+        <Header></Header>
+      <main>
+        <h1 className={heading}>{pageTitle}</h1>
+        {children}
+      </main>
+    </div>
+  )
+}
+
+export default Layout
